Inline src/utils.js into src/raven.js for better compression
